Working on a sript that was given by Phylroy to get data from json

+++ Creating_excel_tables.py
@@ -19,8 +19,9 @@ class TableMaker():
     TOTAL_END_USES_GJ_PER_M2 = None
     TOTAL_END_USES_GJ_PER_YEAR = None
 
-    def __init__(self, desired_values_gj_per_year, desired_values_gj_per_m2, url='C:/Users/thotran/Desktop/simulations.json'):
-        self.TOTAL_END_USES_GJ_PER_M2 =desired_values_gj_per_m2
+    def __init__(self, desired_values_gj_per_year, desired_values_gj_per_m2,
+                 url='C:/Users/thotran/Desktop/simulations.json'):
+        self.TOTAL_END_USES_GJ_PER_M2 = desired_values_gj_per_m2
         self.TOTAL_END_USES_GJ_PER_YEAR = desired_values_gj_per_year
         pd.set_option('display.max_columns', None)
         with open(url) as f:
@@ -34,7 +35,7 @@ class TableMaker():
         self.main_table.drop(self.main_table[(~self.main_table[column].isin(values))].index, inplace=True)
         return self
 
-    def necb_comparison_table(self, columns=[TOTAL_END_USES_GJ_PER_YEAR ,TOTAL_END_USES_GJ_PER_M2],
+    def necb_comparison_table(self, columns=[TOTAL_END_USES_GJ_PER_YEAR, TOTAL_END_USES_GJ_PER_M2],
                               templates=['NECB2015', 'NECB2017']):
         # https://stackoverflow.com/questions/35268817/unique-combinations-of-values-in-selected-columns-in-pandas-data-frame-and-count
         # Create a table just with cities and building types. Keeping 'count' for now for debugging.
@@ -50,39 +51,47 @@ class TableMaker():
                         (self.main_table[self.BUILDING_TYPE] == row[self.BUILDING_TYPE]) &
                         (self.main_table[self.CITY] == row[self.CITY])
                         ].iloc[0][column], axis=1)
-        #print(table)
+        print(table)
         return table
 
-    def creating_tables(gj_per_year, gj_per_m2):
-        tm = TableMaker(gj_per_year, gj_per_m2)
-        #filters by cities you want
-        tm.filter_out_rows_by_values(tm.BUILDING_TYPE, ['SmallOffice',
-                                                        'MediumOffice',
-                                                        'LargeOffice',
-                                                        'RetailStandalone',
-                                                        'RetailStripmall',
-                                                        'MidriseApartment',
-                                                        'HighriseApartment'
-                                                        ])
 
-        #Filters by buildings you want.
-        tm.filter_out_rows_by_values(tm.CITY, ['Abbotsford Intl AP'])#replace with cities
+def creating_tables(gj_per_year, gj_per_m2):
+    tm = TableMaker(gj_per_year, gj_per_m2)
+    # filters by cities you want
+    tm.filter_out_rows_by_values(tm.BUILDING_TYPE, ['SmallOffice',
+                                                    'MediumOffice',
+                                                    'LargeOffice',
+                                                    'RetailStandalone',
+                                                    'RetailStripmall',
+                                                    'MidriseApartment',
+                                                    'HighriseApartment'
+                                                    ])
 
+    # Filters by buildings you want.
+    tm.filter_out_rows_by_values(tm.CITY, ['Abbotsford Intl AP'])  # replace with cities
 
-        # This creates a table with necb201x values in columns rather than rows.
-        # the first arguement is the column from the simulations.json top level down, you can add more as you need.. the second is the vintages you want in the table.
-        # You can have a single vintage if you want to have a table like 4.3.3.a
-        comparison_table = tm.necb_comparison_table(['end_uses_eui.total_end_uses_gj_per_m2'],['NECB2015', 'NECB2017'])
-        # You should be able to rename the columns after the table is created to be pretty.
-        return comparison_table
+    # This creates a table with necb201x values in columns rather than rows.
+    # the first arguement is the column from the simulations.json top level down, you can add more as you need.. the second is the vintages you want in the table.
+    # You can have a single vintage if you want to have a table like 4.3.3.a
+    comparison_table = tm.necb_comparison_table([f"{TOTAL_END_USES_GJ_PER_YEAR}"_, f"{TOTAL_END_USES_GJ_PER_M2}"],
+                                                ['NECB2015', 'NECB2017'])
+    # You should be able to rename the columns after the table is created to be pretty.
+    return comparison_table
+
 
 wanted_values_gj = []
 wanted_values_gj_per_m2 = []
+arr = []
+if (len(wanted_values_gj) != len(wanted_values_gj_per_m2)):
+    raise ('please make sure that the two arrays are equal in length')
+for i in range(0, len(wanted_values_gj)):
+    arr[i] = creating_tables(wanted_values_gj[i],wanted_values_gj_per_m2[i])
 
-output = pd.ExcelWriter('C:/Users/thotran/Desktop/panda_simple.xlsx', engine = 'xlsxwriter')
+raise ('hell')
+# this already works I am testing if the current for loop above will work
+output = pd.ExcelWriter('C:/Users/thotran/Desktop/panda_simple.xlsx', engine='xlsxwriter')
 my_file = Path('C:/Users/thotran/Desktop/panda_simple.xlsx')
 if my_file.is_file():
     os.remove('C:/Users/thotran/Desktop/panda_simple.xlsx')
-comparison_table.to_excel(output, sheet_name = 'NECB2011')
+comparison_table.to_excel(output, sheet_name='NECB2011')
 output.close()
-
